feat(agent): move github commands to input capability
